requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
cffi==1.17.1
cryptography==43.0.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
pytz==2024.2
PyYAML==6.0.2
qq-botpy==1.2.1
setuptools==75.3.0
shiboken6==6.8.0.2
six==1.16.0
tzdata==2024.2
tzlocal==5.2
websockets==13.1
yarl==1.17.1
